@@ -6,18 +6,17 @@
         "start": "./node_modules/.bin/webpack serve --port=9060 --mode=development"
     },
     "devDependencies": {
-        "@babel/core": "^7.12.10",
-        "@babel/preset-env": "^7.12.11",
+        "@babel/core": "^7.13.14",
+        "@babel/preset-env": "^7.13.12",
         "babel-loader": "^8.2.2",
-        "copy-webpack-plugin": "^7.0.0",
-        "css-loader": "^5.0.1",
+        "copy-webpack-plugin": "^8.1.1",
+        "css-loader": "^5.2.0",
         "file-loader": "^6.2.0",
-        "html-loader": "^1.3.2",
+        "html-loader": "^2.1.2",
         "style-loader": "^2.0.0",
-        "terser-webpack-plugin": "^5.0.3",
+        "terser-webpack-plugin": "^5.1.1",
         "url-loader": "^4.1.1",
-        "webpack": "^5.11.1",
-        "webpack-cli": "^4.3.0"
-    },
-    "dependencies": {}
+        "webpack": "^5.30.0",
+        "webpack-cli": "^4.6.0"
+    }
 }
