@@ -1,15 +1,19 @@
-import { Component } from 'react';
+import { useEffect, useState, useRef } from 'react';
 import { FontAwesomeIcon as Icon } from '@fortawesome/react-fontawesome';
-import { faAngleDoubleRight, faSpinner } from '@fortawesome/free-solid-svg-icons';
+import { faStream } from '@fortawesome/free-solid-svg-icons';
 import styled from 'styled-components/macro';
-import { getStatementsBySubject, getStatementsByObjectAndPredicate } from 'services/backend/statements';
 import { getResearchFieldsStats } from 'services/backend/stats';
-import { Link } from 'react-router-dom';
-import { reverse } from 'named-urls';
-import { Button } from 'reactstrap';
-import ROUTES from 'constants/routes.js';
 import { TransitionGroup, CSSTransition } from 'react-transition-group';
-import { PREDICATES, MISC, CLASSES } from 'constants/graphSettings';
+import { MISC, CLASSES } from 'constants/graphSettings';
+import Autocomplete from 'components/Autocomplete/Autocomplete';
+import { resourcesUrl } from 'services/backend/resources';
+import { has } from 'lodash';
+import ContentLoader from 'react-content-loader';
+import PropTypes from 'prop-types';
+import ROUTES from 'constants/routes';
+import Breadcrumbs from 'components/Breadcrumbs/Breadcrumbs';
+import { Link } from 'react-router-dom';
+import { Button } from 'reactstrap';
 import { reverseWithSlug } from 'utils';
 
 /* Bootstrap card column is not working correctly working with vertical alignment,
@@ -22,7 +26,7 @@ const Card = styled.div`
     border: 0 !important;
     border-radius: 12px !important;
     min-height: 85px;
-    flex: 0 0 calc(33% - 20px) !important;
+    flex: 0 0 calc(20% - 20px) !important;
     margin: 10px;
     transition: opacity 0.2s;
     justify-content: center;
@@ -57,19 +61,24 @@ const CardTitle = styled.h5`
     padding: 0 5px;
 `;
 
-const BreadcrumbLink = styled.span`
-    cursor: pointer;
-    margin-right: 5px;
-
-    &:hover {
-        text-decoration: underline;
-    }
-`;
-
 const PaperAmount = styled.div`
     opacity: 0.5;
     font-size: 80%;
     text-align: center;
+`;
+
+const ArrowCards = styled.div`
+    background: transparent;
+    position: absolute;
+    left: 0px;
+    transform: translate(89px);
+    bottom: -15px;
+    width: 0;
+    height: 0;
+    border-left: 15px solid transparent;
+    border-right: 15px solid transparent;
+    border-top: 15px solid #fff;
+    filter: drop-shadow(1px 1px 0px rgba(0, 0, 0, 0.13));
 `;
 
 const AnimationContainer = styled(CSSTransition)`
@@ -86,193 +95,146 @@ const AnimationContainer = styled(CSSTransition)`
     }
 `;
 
-const MAX_PAPER_AMOUNT = 24;
+const ShowMore = styled(Card)`
+    background: ${props => props.theme.light}!important;
+    color: ${props => props.theme.bodyColor} !important;
+    text-align: center;
+`;
 
-class ResearchFieldCards extends Component {
-    state = {
-        researchFields: [],
-        breadcrumb: [],
-        papers: null,
-        error: '',
-        stats: {}
-    };
+const ResearchFieldCards = ({ selectedResearchField, handleFieldSelect, researchFields, isLoading }) => {
+    const [stats, setStats] = useState(null);
+    const rfAutocompleteRef = useRef(null);
+    const [showMoreFields, setShowMoreFields] = useState(false);
 
-    componentDidMount() {
-        this.getFields(MISC.RESEARCH_FIELD_MAIN, 'Main');
-        this.fetchResearchFieldsStats();
-    }
+    useEffect(() => {
+        fetchResearchFieldsStats();
+    }, []);
 
-    fetchResearchFieldsStats = () => {
+    const fetchResearchFieldsStats = () => {
         return getResearchFieldsStats().then(results => {
-            this.setState({ stats: results });
+            setStats(results);
         });
     };
 
-    async getFields(fieldId, label, addBreadcrumb = true) {
-        try {
-            await getStatementsBySubject({ id: fieldId })
-                .then(async res => {
-                    let researchFields = [];
-                    res.forEach(elm => {
-                        researchFields.push({
-                            label: elm.object.label,
-                            id: elm.object.id
-                        });
-                    });
-
-                    // sort research fields alphabetically
-                    researchFields = researchFields.sort((a, b) => {
-                        return a.label.localeCompare(b.label);
-                    });
-
-                    this.setState({
-                        researchFields,
-                        error: ''
-                    });
-
-                    if (addBreadcrumb) {
-                        const breadcrumb = this.state.breadcrumb;
-
-                        breadcrumb.push({
-                            label: label,
-                            id: fieldId
-                        });
-
-                        this.setState({
-                            breadcrumb: breadcrumb
-                        });
-                    }
-
-                    if (fieldId !== MISC.RESEARCH_FIELD_MAIN) {
-                        this.setState({
-                            papers: null // to show loading indicator
-                        });
-
-                        let papers = await getStatementsByObjectAndPredicate({
-                            objectId: fieldId,
-                            predicateId: PREDICATES.HAS_RESEARCH_FIELD,
-                            page: 0,
-                            items: MAX_PAPER_AMOUNT,
-                            sortBy: 'created_at',
-                            desc: true
-                        });
-
-                        papers = papers.filter(statement => statement.subject.classes.includes(CLASSES.PAPER));
-
-                        this.setState({
-                            papers
-                        });
-                    }
-                })
-                .catch(e => {
-                    this.setState({
-                        error: e.message
-                    });
-                });
-        } catch (e) {
-            this.setState({
-                error: e.message
-            });
-        }
-    }
-
-    handleClickBreadcrumb(fieldId, label) {
-        const activeIndex = this.state.breadcrumb.findIndex(breadcrumb => breadcrumb.id === fieldId);
-        const breadcrumb = this.state.breadcrumb.slice(0, activeIndex + 1); //remove the items after the clicked link
-
-        this.setState({
-            breadcrumb
-        });
-
-        this.getFields(fieldId, label, false);
-    }
-
-    render() {
-        if (this.state.error) {
-            return <div className="text-center mt-5 text-danger">{this.state.error}</div>;
-        }
-
-        const showPapers = this.state.breadcrumb.length > 1;
-        const currentField = this.state.breadcrumb[this.state.breadcrumb.length - 1];
-        const researchFieldLink = this.state.breadcrumb.length
-            ? reverseWithSlug(ROUTES.RESEARCH_FIELD, { researchFieldId: currentField.id, slug: currentField.label })
-            : null;
-
-        return (
-            <div className="mt-3">
-                <div className="d-flex justify-content-between align-items-start">
-                    <div>
-                        {this.state.breadcrumb.map((field, index) => (
-                            <BreadcrumbLink key={field.id} onClick={() => this.handleClickBreadcrumb(field.id, field.label)}>
-                                {field.label} {index !== this.state.breadcrumb.length - 1 && <Icon icon={faAngleDoubleRight} />}
-                            </BreadcrumbLink>
-                        ))}
-                    </div>
-                    {currentField && currentField.id !== MISC.RESEARCH_FIELD_MAIN && (
-                        <Button tag={Link} to={researchFieldLink} color="light" size="sm" className="flex-shrink-0">
+    return (
+        <>
+            <div className="row" style={{ position: 'relative' }}>
+                <h1 className="col-md-8 h5 flex-shrink-0 mb-0">
+                    <Icon icon={faStream} className="text-primary" /> Browse by research field
+                </h1>
+                <div className="col-md-4 mt-2 mt-md-0 flex-end d-flex">
+                    {selectedResearchField.id !== MISC.RESEARCH_FIELD_MAIN && (
+                        <Button
+                            tag={Link}
+                            to={reverseWithSlug(ROUTES.RESEARCH_FIELD, {
+                                researchFieldId: selectedResearchField.id,
+                                slug: selectedResearchField.label
+                            })}
+                            color="light"
+                            size="sm"
+                            className="flex-shrink-0 mr-2"
+                        >
                             Visit field page
                         </Button>
                     )}
+                    <Autocomplete
+                        requestUrl={resourcesUrl}
+                        optionsClass={CLASSES.RESEARCH_FIELD}
+                        placeholder="Search for fields..."
+                        onItemSelected={selected => {
+                            // blur the field allows to focus and open the menu again
+                            rfAutocompleteRef.current && rfAutocompleteRef.current.blur();
+                            handleFieldSelect(selected);
+                        }}
+                        value={selectedResearchField.id !== MISC.RESEARCH_FIELD_MAIN ? selectedResearchField : null}
+                        allowCreate={false}
+                        autoLoadOption={true}
+                        cssClasses="form-control-sm"
+                        isDisabled={isLoading}
+                        innerRef={rfAutocompleteRef}
+                    />
                 </div>
-
-                <hr className="mt-3 mb-3" />
-                <div>
-                    <TransitionGroup id="research-field-cards" className="mt-2 justify-content-center d-flex flex-wrap" exit={false}>
-                        {this.state.researchFields.map(field => {
-                            return (
-                                <AnimationContainer key={field.id} classNames="fadeIn" timeout={{ enter: 500, exit: 0 }}>
-                                    <Card
-                                        role="button"
-                                        disabled={this.state.stats[field.id] === 0}
-                                        onClick={() => this.getFields(field.id, field.label)}
-                                    >
-                                        <CardTitle className="card-title m-0 text-center">{field.label}</CardTitle>
-                                        <PaperAmount>{this.state.stats[field.id]} papers</PaperAmount>
-                                    </Card>
-                                </AnimationContainer>
-                            );
-                        })}
-                    </TransitionGroup>
-                </div>
-                {showPapers && (
-                    <div className="mt-3">
-                        <h2 className="h5">
-                            <Link to={researchFieldLink}>{currentField.label}</Link> papers
-                        </h2>
-
-                        {!this.state.papers && (
-                            <div className="mt-5 text-center">
-                                <Icon icon={faSpinner} spin /> Loading
-                            </div>
-                        )}
-
-                        {this.state.papers && this.state.papers.length === 0 ? <div className="mt-5 text-center">No papers found</div> : null}
-
-                        {this.state.papers && (
-                            <>
-                                <ul className="mt-3">
-                                    {this.state.papers.map((paper, index) => (
-                                        <li key={index}>
-                                            <Link to={reverse(ROUTES.VIEW_PAPER, { resourceId: paper.subject.id })}>
-                                                {paper.subject.label ? paper.subject.label : <i>No title</i>}
-                                            </Link>
-                                        </li>
-                                    ))}
-                                </ul>
-                                {this.state.papers.length >= MAX_PAPER_AMOUNT && (
-                                    <div className="text-center">
-                                        <Button tag={Link} to={researchFieldLink} size="sm" color="primary">
-                                            View more papers
-                                        </Button>
-                                    </div>
-                                )}
-                            </>
-                        )}
-                    </div>
-                )}
             </div>
-        );
-    }
-}
+            <hr className="mt-3 mb-1" />
+            {MISC.RESEARCH_FIELD_MAIN !== selectedResearchField.id && (
+                <>
+                    <Breadcrumbs backgroundWhite researchFieldId={selectedResearchField.id} onFieldClick={handleFieldSelect} disableLastField />
+                    <hr className="mt-1 mb-1" />
+                </>
+            )}
+            {!isLoading && stats && researchFields.length > 0 && (
+                <div className="mt-3">
+                    <div>
+                        <TransitionGroup id="research-field-cards" className="mt-2 justify-content-center d-flex flex-wrap" exit={false}>
+                            {researchFields.slice(0, 9).map(field => {
+                                return (
+                                    <AnimationContainer key={field.id} classNames="fadeIn" timeout={{ enter: 500, exit: 0 }}>
+                                        <Card
+                                            role="button"
+                                            disabled={has(stats, field.id) && stats[field.id] === 0}
+                                            onClick={() => handleFieldSelect(field)}
+                                        >
+                                            <CardTitle className="card-title m-0 text-center">{field.label}</CardTitle>
+                                            <PaperAmount>{has(stats, field.id) ? stats[field.id] : 0} papers</PaperAmount>
+                                        </Card>
+                                    </AnimationContainer>
+                                );
+                            })}
+                            {researchFields.length > 9 &&
+                                showMoreFields &&
+                                researchFields.slice(9).map(field => (
+                                    <AnimationContainer key={field.id} classNames="fadeIn" timeout={{ enter: 500, exit: 0 }}>
+                                        <Card
+                                            role="button"
+                                            disabled={has(stats, field.id) && stats[field.id] === 0}
+                                            onClick={() => handleFieldSelect(field)}
+                                        >
+                                            <CardTitle className="card-title m-0 text-center">{field.label}</CardTitle>
+                                            <PaperAmount>{has(stats, field.id) ? stats[field.id] : 0} papers</PaperAmount>
+                                        </Card>
+                                    </AnimationContainer>
+                                ))}
+                            {researchFields.length > 9 && (
+                                <AnimationContainer classNames="fadeIn" timeout={{ enter: 500, exit: 0 }}>
+                                    <ShowMore role="button" onClick={() => setShowMoreFields(v => !v)}>
+                                        {showMoreFields ? 'Show less fields' : 'Show more fields...'}
+                                    </ShowMore>
+                                </AnimationContainer>
+                            )}
+                        </TransitionGroup>
+                    </div>
+                </div>
+            )}
+            {selectedResearchField.id !== MISC.RESEARCH_FIELD_MAIN && <ArrowCards />}
+            {isLoading && (
+                <div className="mt-3">
+                    <div>
+                        <ContentLoader
+                            height="10%"
+                            width="100%"
+                            viewBox="0 0 100 10"
+                            style={{ width: '100% !important' }}
+                            backgroundColor="#f3f3f3"
+                            foregroundColor="#ecebeb"
+                        >
+                            <rect x="2" y="0" rx="2" ry="2" width="15" height="7" />
+                            <rect x="22" y="0" rx="2" ry="2" width="15" height="7" />
+                            <rect x="42" y="0" rx="2" ry="2" width="15" height="7" />
+                            <rect x="62" y="0" rx="2" ry="2" width="15" height="7" />
+                            <rect x="82" y="0" rx="2" ry="2" width="15" height="7" />
+                        </ContentLoader>
+                    </div>
+                </div>
+            )}
+        </>
+    );
+};
+
+ResearchFieldCards.propTypes = {
+    selectedResearchField: PropTypes.object,
+    researchFields: PropTypes.array,
+    handleFieldSelect: PropTypes.func,
+    isLoading: PropTypes.bool.isRequired
+};
 
 export default ResearchFieldCards;
