@@ -1,14 +1,15 @@
-import { Component } from 'react';
-import { Row, Col } from 'reactstrap';
+import { useEffect, useState } from 'react';
 import { Link } from 'react-router-dom';
 import { reverse } from 'named-urls';
 import styled from 'styled-components';
 import { FontAwesomeIcon as Icon } from '@fortawesome/react-fontawesome';
 import { faCalendar } from '@fortawesome/free-solid-svg-icons';
 import ROUTES from 'constants/routes.js';
+import RelativeBreadcrumbs from 'components/RelativeBreadcrumbs/RelativeBreadcrumbs';
+import UserAvatar from 'components/UserAvatar/UserAvatar';
 import Authors from './Authors';
 import PropTypes from 'prop-types';
-import { getPublicationMonth, getPublicationYear, getAuthors } from 'utils';
+import { getPublicationMonth, getPublicationYear, getAuthors, getResearchField } from 'utils';
 import moment from 'moment';
 import ContentLoader from 'react-content-loader';
 
@@ -27,105 +28,97 @@ const PaperCardStyled = styled.div`
     }
 `;
 
-class PaperCardDynamic extends Component {
-    constructor(props) {
-        super();
-        this.state = { isLoading: true, paperStatements: [], optimizePaperObject: null };
-    }
+const PaperCardDynamic = props => {
+    const [isLoading, setIsLoading] = useState(true);
+    const [optimizedPaperObject, setOptimizePaperObject] = useState(null);
 
-    componentDidMount() {
-        /** this is experimental code, where each paper request its own data when paper Card is mounted **/
-        // getStatementsBySubject({ id: this.props.paper.id }).then(paperStatements => {
-        //     const optimizedPaperObject = getPaperDataForViewAllPapers(paperStatements);
-        //     this.setState({ optimizedPaperObject: optimizedPaperObject, isLoading: false });
-        // });
-
-        if (this.props.paper.paperData && this.props.paper.paperData.statements) {
-            if (this.state.optimizePaperObject === null && this.state.isLoading) {
-                const optimizedPaperObject = this.getPaperDataForViewAllPapers(this.props.paper.paperData.statements);
-                this.setState({ optimizedPaperObject: optimizedPaperObject, isLoading: false });
+    useEffect(() => {
+        setIsLoading(true);
+        if (props.paper.paperData && props.paper.paperData.statements) {
+            if (optimizedPaperObject === null) {
+                setOptimizePaperObject(getPaperDataForViewAllPapers(props.paper.paperData.statements));
+                setIsLoading(false);
             }
         }
-    }
-
-    componentDidUpdate = () => {
-        if (this.props.paper.paperData && this.props.paper.paperData.statements) {
-            if (this.state.optimizePaperObject === null && this.state.isLoading) {
-                const optimizedPaperObject = this.getPaperDataForViewAllPapers(this.props.paper.paperData.statements);
-                this.setState({ optimizedPaperObject: optimizedPaperObject, isLoading: false });
-            }
-        } else {
-            // just make sure that when from outside the props are removed the loading indicator is ensured!
-            if (this.state.optimizedPaperObject !== null) {
-                this.setState({ optimizedPaperObject: null, isLoading: true });
-            }
-        }
-    };
+        // eslint-disable-next-line react-hooks/exhaustive-deps
+    }, [props.paper.paperData]);
 
     /**
      * Parse paper statements and return an OPTIMIZED paper object for View all papers
      *
      * @param {Array} paperStatements
      */
-    getPaperDataForViewAllPapers = paperStatements => {
-        // research field
-        // publication year
+    const getPaperDataForViewAllPapers = paperStatements => {
         const publicationYear = getPublicationYear(paperStatements)[0]; // gets year[0] and resourceId[1]
         const publicationMonth = getPublicationMonth(paperStatements)[0]; // gets month[0] and resourceId[1]
         const authors = getAuthors(paperStatements);
+        const researchField = getResearchField(paperStatements);
         return {
             publicationYear,
             publicationMonth,
-            authorNames: authors.sort((a, b) => a.created_at.localeCompare(b.created_at))
+            authorNames: authors.sort((a, b) => a.created_at.localeCompare(b.created_at)),
+            researchField
         };
     };
 
-    render() {
-        return (
-            <PaperCardStyled className="list-group-item list-group-item-action">
-                <Row>
-                    <Col sm={12}>
-                        <Link to={reverse(ROUTES.VIEW_PAPER, { resourceId: this.props.paper.id })}>
-                            {this.props.paper.title ? this.props.paper.title : <em>No title</em>}
-                        </Link>
-                        <br />
-                        {!this.state.isLoading && (
+    return (
+        <PaperCardStyled className="list-group-item list-group-item-action">
+            <div className="row">
+                <div className="col-9">
+                    <Link to={reverse(ROUTES.VIEW_PAPER, { resourceId: props.paper.id })}>
+                        {props.paper.title ? props.paper.title : <em>No title</em>}
+                    </Link>
+                    <br />
+                    {!isLoading && (
+                        <>
                             <small>
-                                <Authors authors={this.state.optimizedPaperObject.authorNames} />
-                                {(this.state.optimizedPaperObject.publicationMonth || this.state.optimizedPaperObject.publicationYear) && (
+                                <Authors authors={optimizedPaperObject.authorNames} />
+                                {(optimizedPaperObject.publicationMonth || optimizedPaperObject.publicationYear) && (
                                     <Icon size="sm" icon={faCalendar} className="ml-2 mr-1" />
                                 )}
-                                {this.state.optimizedPaperObject.publicationMonth && this.state.optimizedPaperObject.publicationMonth > 0
-                                    ? moment(this.state.optimizedPaperObject.publicationMonth, 'M').format('MMMM')
+                                {optimizedPaperObject.publicationMonth && optimizedPaperObject.publicationMonth > 0
+                                    ? moment(optimizedPaperObject.publicationMonth, 'M').format('MMMM')
                                     : ''}{' '}
-                                {this.state.optimizedPaperObject.publicationYear}
+                                {optimizedPaperObject.publicationYear}
                             </small>
-                        )}
-                        {/*Show Loading Dynamic data indicator if we are loading */}
-                        {this.state.isLoading && (
-                            <div style={{ display: 'ruby' }}>
-                                <span>Loading</span>
-                                <ContentLoader
-                                    style={{ marginTop: '-8px', width: '8% !important' }}
-                                    height={30}
-                                    width="8%"
-                                    viewBox="0 0 100 30"
-                                    speed={2}
-                                    backgroundColor="#f3f3f3"
-                                    foregroundColor="#ccc"
-                                >
-                                    <circle cx="6" cy="18" r="4" />
-                                    <circle cx="16" cy="18" r="4" />
-                                    <circle cx="26" cy="18" r="4" />
-                                </ContentLoader>
+                            <div className="d-block d-md-none">
+                                <RelativeBreadcrumbs researchField={optimizedPaperObject.researchField} />
                             </div>
-                        )}
-                    </Col>
-                </Row>
-            </PaperCardStyled>
-        );
-    }
-}
+                        </>
+                    )}
+
+                    {/*Show Loading Dynamic data indicator if we are loading */}
+                    {isLoading && (
+                        <div style={{ display: 'ruby' }}>
+                            <span>Loading</span>
+                            <ContentLoader
+                                style={{ marginTop: '-8px', width: '8% !important' }}
+                                height={30}
+                                width="8%"
+                                viewBox="0 0 100 30"
+                                speed={2}
+                                backgroundColor="#f3f3f3"
+                                foregroundColor="#ccc"
+                            >
+                                <circle cx="6" cy="18" r="4" />
+                                <circle cx="16" cy="18" r="4" />
+                                <circle cx="26" cy="18" r="4" />
+                            </ContentLoader>
+                        </div>
+                    )}
+                </div>
+                {!isLoading && (
+                    <div className="col-3 text-right d-flex align-items-end" style={{ flexDirection: 'column' }}>
+                        <div style={{ flex: 1 }} className="d-none d-md-block">
+                            <RelativeBreadcrumbs researchField={optimizedPaperObject.researchField} />
+                        </div>
+                        <UserAvatar userId={props.paper?.created_by} />
+                    </div>
+                )}
+            </div>
+        </PaperCardStyled>
+    );
+};
 
 PaperCardDynamic.propTypes = {
     paper: PropTypes.shape({
@@ -134,7 +127,12 @@ PaperCardDynamic.propTypes = {
         authorNames: PropTypes.array,
         publicationMonth: PropTypes.string,
         publicationYear: PropTypes.string,
-        paperData: PropTypes.object
+        paperData: PropTypes.object,
+        researchField: PropTypes.shape({
+            id: PropTypes.string.isRequired,
+            label: PropTypes.string
+        }),
+        created_by: PropTypes.string
     }).isRequired
 };
 
