@@ -1,1 +1,1 @@
-!function(){var e={240:function(e,n,t){"use strict";var r=t(645),o=t.n(r)()((function(e){return e[1]}));o.push([e.id,".orkg-widget-box {\n    font-family: 'Helvetica Neue', Helvetica, Arial, sans-serif;\n    border: 1px solid rgb(232, 97, 97);\n}\n.orkg-widget-link {\n    color: #fff;\n    background-color: rgb(232, 97, 97);\n    border-color: rgb(232, 97, 97);\n    padding: 6px 12px;\n    display: block;\n    margin-bottom: 0 !important;\n    font-size: 14px;\n    font-weight: 400;\n    line-height: 1.42857143;\n    text-align: center;\n    white-space: nowrap;\n    touch-action: manipulation;\n    cursor: pointer;\n    user-select: none;\n    background-image: none;\n    border: 1px solid transparent;\n    text-decoration: none;\n}\n.orkg-widget-link:hover,\n.orkg-widget-link:focus {\n    color: #fff;\n    background-color: #e34040;\n    border-color: #e23434;\n}\n.orkg-widget-description {\n    padding: 5px;\n}\n.orkg-widget-label {\n    position: relative;\n    left: -3px;\n    text-align: left;\n    display: inline-block;\n    background: transparent;\n}\n.orkg-widget-icon {\n    width: 24px;\n    vertical-align: middle;\n}\n.orkg-widget-statements {\n    float: right;\n    font-weight: bold;\n}\n",""]),n.Z=o},645:function(e){"use strict";e.exports=function(e){var n=[];return n.toString=function(){return this.map((function(n){var t=e(n);return n[2]?"@media ".concat(n[2]," {").concat(t,"}"):t})).join("")},n.i=function(e,t,r){"string"==typeof e&&(e=[[null,e,""]]);var o={};if(r)for(var i=0;i<this.length;i++){var a=this[i][0];null!=a&&(o[a]=!0)}for(var s=0;s<e.length;s++){var c=[].concat(e[s]);r&&o[c[0]]||(t&&(c[2]?c[2]="".concat(t," and ").concat(c[2]):c[2]=t),n.push(c))}},n}},440:function(e){e.exports='<div class="orkg-widget-box"> <a href="#" class="orkg-widget-link" target="_blank"> <span class="orkg-widget-label"> <img src="#" class="orkg-widget-icon"/> <span class="orkg-widget-txt-link">Open in ORKG</span> </span> </a> <div class="orkg-widget-description"> <span class="orkg-widget-text-statements">Number of statements</span> <span class="orkg-widget-statements">0</span> </div> </div> '},379:function(e,n,t){"use strict";var r,o=function(){return void 0===r&&(r=Boolean(window&&document&&document.all&&!window.atob)),r},i=function(){var e={};return function(n){if(void 0===e[n]){var t=document.querySelector(n);if(window.HTMLIFrameElement&&t instanceof window.HTMLIFrameElement)try{t=t.contentDocument.head}catch(e){t=null}e[n]=t}return e[n]}}(),a=[];function s(e){for(var n=-1,t=0;t<a.length;t++)if(a[t].identifier===e){n=t;break}return n}function c(e,n){for(var t={},r=[],o=0;o<e.length;o++){var i=e[o],c=n.base?i[0]+n.base:i[0],d=t[c]||0,l="".concat(c," ").concat(d);t[c]=d+1;var u=s(l),f={css:i[1],media:i[2],sourceMap:i[3]};-1!==u?(a[u].references++,a[u].updater(f)):a.push({identifier:l,updater:h(f,n),references:1}),r.push(l)}return r}function d(e){var n=document.createElement("style"),r=e.attributes||{};if(void 0===r.nonce){var o=t.nc;o&&(r.nonce=o)}if(Object.keys(r).forEach((function(e){n.setAttribute(e,r[e])})),"function"==typeof e.insert)e.insert(n);else{var a=i(e.insert||"head");if(!a)throw new Error("Couldn't find a style target. This probably means that the value for the 'insert' parameter is invalid.");a.appendChild(n)}return n}var l,u=(l=[],function(e,n){return l[e]=n,l.filter(Boolean).join("\n")});function f(e,n,t,r){var o=t?"":r.media?"@media ".concat(r.media," {").concat(r.css,"}"):r.css;if(e.styleSheet)e.styleSheet.cssText=u(n,o);else{var i=document.createTextNode(o),a=e.childNodes;a[n]&&e.removeChild(a[n]),a.length?e.insertBefore(i,a[n]):e.appendChild(i)}}function g(e,n,t){var r=t.css,o=t.media,i=t.sourceMap;if(o?e.setAttribute("media",o):e.removeAttribute("media"),i&&"undefined"!=typeof btoa&&(r+="\n/*# sourceMappingURL=data:application/json;base64,".concat(btoa(unescape(encodeURIComponent(JSON.stringify(i))))," */")),e.styleSheet)e.styleSheet.cssText=r;else{for(;e.firstChild;)e.removeChild(e.firstChild);e.appendChild(document.createTextNode(r))}}var p=null,m=0;function h(e,n){var t,r,o;if(n.singleton){var i=m++;t=p||(p=d(n)),r=f.bind(null,t,i,!1),o=f.bind(null,t,i,!0)}else t=d(n),r=g.bind(null,t,n),o=function(){!function(e){if(null===e.parentNode)return!1;e.parentNode.removeChild(e)}(t)};return r(e),function(n){if(n){if(n.css===e.css&&n.media===e.media&&n.sourceMap===e.sourceMap)return;r(e=n)}else o()}}e.exports=function(e,n){(n=n||{}).singleton||"boolean"==typeof n.singleton||(n.singleton=o());var t=c(e=e||[],n);return function(e){if(e=e||[],"[object Array]"===Object.prototype.toString.call(e)){for(var r=0;r<t.length;r++){var o=s(t[r]);a[o].references--}for(var i=c(e,n),d=0;d<t.length;d++){var l=s(t[d]);0===a[l].references&&(a[l].updater(),a.splice(l,1))}t=i}}}}},n={};function t(r){if(n[r])return n[r].exports;var o=n[r]={id:r,exports:{}};return e[r](o,o.exports,t),o.exports}t.n=function(e){var n=e&&e.__esModule?function(){return e.default}:function(){return e};return t.d(n,{a:n}),n},t.d=function(e,n){for(var r in n)t.o(n,r)&&!t.o(e,r)&&Object.defineProperty(e,r,{enumerable:!0,get:n[r]})},t.o=function(e,n){return Object.prototype.hasOwnProperty.call(e,n)},function(){"use strict";var e=t(440),n=t.n(e),r=t(379),o=t.n(r),i=t(240),a={insert:"head",singleton:!1},s=(o()(i.Z,a),i.Z.locals,{add:{de:"Artikel zu ORKG hinzufügen",en:"Add paper to ORKG"},open:{de:"In ORKG öffnen",en:"Open in ORKG"},numStatements:{de:"Anzahl der Aussagen",en:"Number of statements"}});function c(e){for(var t=document.getElementsByClassName("orkg-widget"),r=function(r){var o=document.createElement("div");o.innerHTML=n(),o.getElementsByClassName("orkg-widget-icon")[0].src="data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAACAAAAAkCAMAAADfNcjQAAAArlBMVEUAAAD///////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////8tivQqAAAAOXRSTlMABAf6EfcL1oE7xDYYWu/ib5R8VElA5si4l4+JdBzrv08z8d7OsqydhGNeRjArJiHboHgVD6i6Z6J3b8ivAAACPElEQVQ4y3VUB5KjQAwUOWewiY7rnNfe2+v/f+zEDK7yVdmiCmZoqdXSCOjFFMosRDt+vjeFfoHxCC0pH3APWkD6Gs4nDhcNQ4r19Z5ACRDqjNMcxXuKY2PLVOW6J1J4Ofg9zqeLrqQ/sSIdnEh9DfRKDcDeBlCR5B576nxXr9w/rUnkwFgnm+nc0gxtpD8Er53Os9xxcnti+TS29KEEDWUe2R27OA1J042Kwm9SFY46jRfW6QZMGp/MVGpUyb3SuJZ5dccz6ZJGwKilgykZ1IlLe0tKz7H0eeFvQ8DNht7u0bLkGSl8BZiYoojzD6DZQqw+CjnoOFaZi2JYQxtodgWmR15OkVIfa/PNrPE8Ao7cAIUTXG5XlYTHgQ7FFq6AJUkc/LLauS/3rGxLNqYSHkg0GFjTkDW5EaVwXx3u0AyUpIjNX5dvrdTw1FnYwKLwxa4JTa5iJ6uQeArDy749O+oblsHra56BdxL2XOBrTvSYoCbSFxvRKA9GJ+B7bQCJ2p+EoxknatAJ4gvCM8NmHAHX2ZCrQkLhIF51nDt1scXaav05k2dtSbjxcfehm3CU7QCtdES0MBppZCwvwvkbgHGM1p6at7n+nGPNIhvGqqqmQcIjB1E4ndLyTxVnOb/DjEMnYFsdACxNmdw858nXAoB7ELNU7HeZQ/tV3Q3igp5I7brHUMzT4mQY1abv3AsoBSvmEp7oWRuT8vbrLlD2/fOt+6f/wwYx57bsj78QdYUfO5p+xJlkGxrV//A/nMo2F7Pql/EAAAAASUVORK5CYII=";var i=t[r],a=i.getAttribute("data-doi");(function(e){var n="https://www.orkg.org/orkg/api/widgets/?doi="+e;return new Promise((function(e,t){fetch(n).then((function(n){if(n.ok){var r=n.json();if(!r.then)return e(r);r.then(e).catch(t)}else t({error:new Error("Error response. (".concat(n.status,") ").concat(n.statusText)),statusCode:n.status,statusText:n.statusText})})).catch(t)}))})(a).then((function(n){for(o.getElementsByClassName("orkg-widget-txt-link")[0].textContent=s.open[e.language],o.getElementsByClassName("orkg-widget-text-statements")[0].textContent=s.numStatements[e.language],o.getElementsByClassName("orkg-widget-statements")[0].textContent=n.num_statements,o.getElementsByClassName("orkg-widget-link")[0].href="https://www.orkg.org/orkg/paper/"+n.id;o.children.length>0;)i.appendChild(o.children[0])})).catch((function(n){o.getElementsByClassName("orkg-widget-txt-link")[0].textContent=s.add[e.language],o.getElementsByClassName("orkg-widget-link")[0].href="https://www.orkg.org/orkg/add-paper?entry="+a;var t=o.getElementsByClassName("orkg-widget-description")[0];for(t.parentNode.removeChild(t);o.children.length>0;)i.appendChild(o.children[0])}))},o=0;o<t.length;o++)r(o)}var d=["paper"];function l(e,n){if(!e)throw Error("API method required");if(e=e.toLowerCase(),-1===d.indexOf(e))throw Error("Method ".concat(e," is not supported"));switch(e){case"paper":c(n);break;default:console.warn("No handler defined for ".concat(e))}}!function(e){var n=e[e["ORKG-Widget"]],t=n.q;if(t)for(var r=0;r<t.length;r++)l(t[r][0],t[r][1]);n=l}(window)}()}();
+!function(){"use strict";var e={240:function(e,n,t){var r=t(645),o=t.n(r)()((function(e){return e[1]}));o.push([e.id,".orkg-widget-box {\n    font-family: 'Helvetica Neue', Helvetica, Arial, sans-serif;\n    border: 1px solid rgb(232, 97, 97);\n}\n.orkg-widget-link {\n    color: #fff;\n    background-color: rgb(232, 97, 97);\n    border-color: rgb(232, 97, 97);\n    padding: 6px 12px;\n    display: block;\n    margin-bottom: 0 !important;\n    font-size: 14px;\n    font-weight: 400;\n    line-height: 1.42857143;\n    text-align: center;\n    white-space: nowrap;\n    touch-action: manipulation;\n    cursor: pointer;\n    user-select: none;\n    background-image: none;\n    border: 1px solid transparent;\n    text-decoration: none;\n}\n.orkg-widget-link:hover,\n.orkg-widget-link:focus {\n    color: #fff;\n    background-color: #e34040;\n    border-color: #e23434;\n}\n.orkg-widget-description {\n    padding: 5px;\n}\n.orkg-widget-label {\n    position: relative;\n    left: -3px;\n    text-align: left;\n    display: inline-block;\n    background: transparent;\n}\n.orkg-widget-icon {\n    width: 24px;\n    vertical-align: middle;\n}\n.orkg-widget-statements {\n    float: right;\n    font-weight: bold;\n}\n",""]),n.Z=o},645:function(e){e.exports=function(e){var n=[];return n.toString=function(){return this.map((function(n){var t=e(n);return n[2]?"@media ".concat(n[2]," {").concat(t,"}"):t})).join("")},n.i=function(e,t,r){"string"==typeof e&&(e=[[null,e,""]]);var o={};if(r)for(var i=0;i<this.length;i++){var a=this[i][0];null!=a&&(o[a]=!0)}for(var s=0;s<e.length;s++){var c=[].concat(e[s]);r&&o[c[0]]||(t&&(c[2]?c[2]="".concat(t," and ").concat(c[2]):c[2]=t),n.push(c))}},n}},379:function(e,n,t){var r,o=function(){return void 0===r&&(r=Boolean(window&&document&&document.all&&!window.atob)),r},i=function(){var e={};return function(n){if(void 0===e[n]){var t=document.querySelector(n);if(window.HTMLIFrameElement&&t instanceof window.HTMLIFrameElement)try{t=t.contentDocument.head}catch(e){t=null}e[n]=t}return e[n]}}(),a=[];function s(e){for(var n=-1,t=0;t<a.length;t++)if(a[t].identifier===e){n=t;break}return n}function c(e,n){for(var t={},r=[],o=0;o<e.length;o++){var i=e[o],c=n.base?i[0]+n.base:i[0],d=t[c]||0,l="".concat(c," ").concat(d);t[c]=d+1;var u=s(l),f={css:i[1],media:i[2],sourceMap:i[3]};-1!==u?(a[u].references++,a[u].updater(f)):a.push({identifier:l,updater:h(f,n),references:1}),r.push(l)}return r}function d(e){var n=document.createElement("style"),r=e.attributes||{};if(void 0===r.nonce){var o=t.nc;o&&(r.nonce=o)}if(Object.keys(r).forEach((function(e){n.setAttribute(e,r[e])})),"function"==typeof e.insert)e.insert(n);else{var a=i(e.insert||"head");if(!a)throw new Error("Couldn't find a style target. This probably means that the value for the 'insert' parameter is invalid.");a.appendChild(n)}return n}var l,u=(l=[],function(e,n){return l[e]=n,l.filter(Boolean).join("\n")});function f(e,n,t,r){var o=t?"":r.media?"@media ".concat(r.media," {").concat(r.css,"}"):r.css;if(e.styleSheet)e.styleSheet.cssText=u(n,o);else{var i=document.createTextNode(o),a=e.childNodes;a[n]&&e.removeChild(a[n]),a.length?e.insertBefore(i,a[n]):e.appendChild(i)}}function g(e,n,t){var r=t.css,o=t.media,i=t.sourceMap;if(o?e.setAttribute("media",o):e.removeAttribute("media"),i&&"undefined"!=typeof btoa&&(r+="\n/*# sourceMappingURL=data:application/json;base64,".concat(btoa(unescape(encodeURIComponent(JSON.stringify(i))))," */")),e.styleSheet)e.styleSheet.cssText=r;else{for(;e.firstChild;)e.removeChild(e.firstChild);e.appendChild(document.createTextNode(r))}}var p=null,m=0;function h(e,n){var t,r,o;if(n.singleton){var i=m++;t=p||(p=d(n)),r=f.bind(null,t,i,!1),o=f.bind(null,t,i,!0)}else t=d(n),r=g.bind(null,t,n),o=function(){!function(e){if(null===e.parentNode)return!1;e.parentNode.removeChild(e)}(t)};return r(e),function(n){if(n){if(n.css===e.css&&n.media===e.media&&n.sourceMap===e.sourceMap)return;r(e=n)}else o()}}e.exports=function(e,n){(n=n||{}).singleton||"boolean"==typeof n.singleton||(n.singleton=o());var t=c(e=e||[],n);return function(e){if(e=e||[],"[object Array]"===Object.prototype.toString.call(e)){for(var r=0;r<t.length;r++){var o=s(t[r]);a[o].references--}for(var i=c(e,n),d=0;d<t.length;d++){var l=s(t[d]);0===a[l].references&&(a[l].updater(),a.splice(l,1))}t=i}}}}},n={};function t(r){var o=n[r];if(void 0!==o)return o.exports;var i=n[r]={id:r,exports:{}};return e[r](i,i.exports,t),i.exports}t.n=function(e){var n=e&&e.__esModule?function(){return e.default}:function(){return e};return t.d(n,{a:n}),n},t.d=function(e,n){for(var r in n)t.o(n,r)&&!t.o(e,r)&&Object.defineProperty(e,r,{enumerable:!0,get:n[r]})},t.o=function(e,n){return Object.prototype.hasOwnProperty.call(e,n)},function(){var e=t(379),n=t.n(e),r=t(240),o={insert:"head",singleton:!1},i=(n()(r.Z,o),r.Z.locals,{add:{de:"Artikel zu ORKG hinzufügen",en:"Add paper to ORKG"},open:{de:"In ORKG öffnen",en:"Open in ORKG"},numStatements:{de:"Anzahl der Aussagen",en:"Number of statements"}});function a(e){for(var n=document.getElementsByClassName("orkg-widget"),t=function(t){var r=document.createElement("div");r.innerHTML='<div class="orkg-widget-box"> <a href="#" class="orkg-widget-link" target="_blank"> <span class="orkg-widget-label"> <img src="#" class="orkg-widget-icon"/> <span class="orkg-widget-txt-link">Open in ORKG</span> </span> </a> <div class="orkg-widget-description"> <span class="orkg-widget-text-statements">Number of statements</span> <span class="orkg-widget-statements">0</span> </div> </div> ',r.getElementsByClassName("orkg-widget-icon")[0].src="data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAACAAAAAkCAMAAADfNcjQAAAArlBMVEUAAAD///////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////8tivQqAAAAOXRSTlMABAf6EfcL1oE7xDYYWu/ib5R8VElA5si4l4+JdBzrv08z8d7OsqydhGNeRjArJiHboHgVD6i6Z6J3b8ivAAACPElEQVQ4y3VUB5KjQAwUOWewiY7rnNfe2+v/f+zEDK7yVdmiCmZoqdXSCOjFFMosRDt+vjeFfoHxCC0pH3APWkD6Gs4nDhcNQ4r19Z5ACRDqjNMcxXuKY2PLVOW6J1J4Ofg9zqeLrqQ/sSIdnEh9DfRKDcDeBlCR5B576nxXr9w/rUnkwFgnm+nc0gxtpD8Er53Os9xxcnti+TS29KEEDWUe2R27OA1J042Kwm9SFY46jRfW6QZMGp/MVGpUyb3SuJZ5dccz6ZJGwKilgykZ1IlLe0tKz7H0eeFvQ8DNht7u0bLkGSl8BZiYoojzD6DZQqw+CjnoOFaZi2JYQxtodgWmR15OkVIfa/PNrPE8Ao7cAIUTXG5XlYTHgQ7FFq6AJUkc/LLauS/3rGxLNqYSHkg0GFjTkDW5EaVwXx3u0AyUpIjNX5dvrdTw1FnYwKLwxa4JTa5iJ6uQeArDy749O+oblsHra56BdxL2XOBrTvSYoCbSFxvRKA9GJ+B7bQCJ2p+EoxknatAJ4gvCM8NmHAHX2ZCrQkLhIF51nDt1scXaav05k2dtSbjxcfehm3CU7QCtdES0MBppZCwvwvkbgHGM1p6at7n+nGPNIhvGqqqmQcIjB1E4ndLyTxVnOb/DjEMnYFsdACxNmdw858nXAoB7ELNU7HeZQ/tV3Q3igp5I7brHUMzT4mQY1abv3AsoBSvmEp7oWRuT8vbrLlD2/fOt+6f/wwYx57bsj78QdYUfO5p+xJlkGxrV//A/nMo2F7Pql/EAAAAASUVORK5CYII=";var o=n[t],a=o.getAttribute("data-doi");(function(e){var n="https://www.orkg.org/orkg/api/widgets/?doi="+e;return new Promise((function(e,t){fetch(n).then((function(n){if(n.ok){var r=n.json();if(!r.then)return e(r);r.then(e).catch(t)}else t({error:new Error("Error response. (".concat(n.status,") ").concat(n.statusText)),statusCode:n.status,statusText:n.statusText})})).catch(t)}))})(a).then((function(n){for(r.getElementsByClassName("orkg-widget-txt-link")[0].textContent=i.open[e.language],r.getElementsByClassName("orkg-widget-text-statements")[0].textContent=i.numStatements[e.language],r.getElementsByClassName("orkg-widget-statements")[0].textContent=n.num_statements,r.getElementsByClassName("orkg-widget-link")[0].href="https://www.orkg.org/orkg/paper/"+n.id;r.children.length>0;)o.appendChild(r.children[0])})).catch((function(n){r.getElementsByClassName("orkg-widget-txt-link")[0].textContent=i.add[e.language],r.getElementsByClassName("orkg-widget-link")[0].href="https://www.orkg.org/orkg/add-paper?entry="+a;var t=r.getElementsByClassName("orkg-widget-description")[0];for(t.parentNode.removeChild(t);r.children.length>0;)o.appendChild(r.children[0])}))},r=0;r<n.length;r++)t(r)}var s=["paper"];function c(e,n){if(!e)throw Error("API method required");if(e=e.toLowerCase(),-1===s.indexOf(e))throw Error("Method ".concat(e," is not supported"));switch(e){case"paper":a(n);break;default:console.warn("No handler defined for ".concat(e))}}!function(e){var n=e[e["ORKG-Widget"]],t=n.q;if(t)for(var r=0;r<t.length;r++)c(t[r][0],t[r][1]);n=c}(window)}()}();
