@@ -73,7 +73,7 @@ const PropertyValue = ({ id, label, property, similar, filterControlData, update
         <>
             <Button onClick={handleOpenStatementBrowser} color="link" className="text-left text-light m-0 p-0">
                 <DescriptionTooltip
-                    id={property.id}
+                    id={property?.id}
                     typeId={PREDICATE_TYPE_ID}
                     extraContent={similar && similar.length ? `This property is merged with : ${similar.join(', ')}` : ''}
                 >
