@@ -1,6 +1,6 @@
 import { useState, useEffect, useCallback } from 'react';
 import { getStatementsBySubject, getStatementsBySubjectAndPredicate, getStatementsByObjectAndPredicate } from 'services/backend/statements';
-import { getUserInformationById } from 'services/backend/users';
+import { getContributorInformationById } from 'services/backend/contributors';
 import { getObservatoryAndOrganizationInformation } from 'services/backend/observatories';
 import { getResource } from 'services/backend/resources';
 import { getComparison } from 'services/similarity/index';
@@ -148,6 +148,8 @@ function useComparison() {
                         const resources = filterObjectOfStatementsByPredicate(statements, PREDICATES.RELATED_RESOURCES, false);
                         const figures = filterObjectOfStatementsByPredicate(statements, PREDICATES.RELATED_FIGURE, false);
                         const visualizations = filterObjectOfStatementsByPredicate(statements, PREDICATES.HAS_VISUALIZATION, false);
+                        const subject = filterObjectOfStatementsByPredicate(statements, PREDICATES.HAS_SUBJECT, true);
+
                         // Load authors
                         let creators = filterObjectOfStatementsByPredicate(statements, PREDICATES.HAS_AUTHOR, false);
                         if (creators) {
@@ -166,8 +168,10 @@ function useComparison() {
                             hasPreviousVersion: hasPreviousVersion,
                             resources: resources ? resources : [],
                             figures: figures ? figures : [],
-                            visualizations: visualizations ? visualizations : []
+                            visualizations: visualizations ? visualizations : [],
+                            subject: subject
                         });
+
                         // TODO: replace this with ordered feature
                         // Load comparison config
                         const url = filterObjectOfStatementsByPredicate(statements, PREDICATES.URL, true);
@@ -247,7 +251,7 @@ function useComparison() {
     const loadCreatedBy = created_by => {
         // Get Provenance data
         if (created_by && created_by !== MISC.UNKNOWN_ID) {
-            getUserInformationById(created_by)
+            getContributorInformationById(created_by)
                 .then(creator => {
                     setCreatedBy(creator);
                 })
@@ -415,7 +419,7 @@ function useComparison() {
                     subjectId: comparisonData.contributions[0].paperId,
                     predicateId: PREDICATES.HAS_RESEARCH_FIELD
                 }).then(s => {
-                    if (s.length) {
+                    if (s.length && !metaData?.subject) {
                         setResearchField(s[0].object);
                     }
                     return Promise.resolve(comparisonData);
@@ -452,6 +456,7 @@ function useComparison() {
                 setIsLoadingComparisonResult(false);
                 setIsFailedLoadingComparisonResult(true);
             });
+        // eslint-disable-next-line react-hooks/exhaustive-deps
     }, [comparisonType, contributionsList, extendAndSortProperties, responseHash]);
 
     /**
