@@ -8,7 +8,10 @@ import { faCalendar, faFile } from '@fortawesome/free-solid-svg-icons';
 import ROUTES from 'constants/routes.js';
 import PropTypes from 'prop-types';
 import moment from 'moment';
+import UserAvatar from 'components/UserAvatar/UserAvatar';
+import RelativeBreadcrumbs from 'components/RelativeBreadcrumbs/RelativeBreadcrumbs';
 import { getStatementsBySubjects } from 'services/backend/statements';
+import { truncate } from 'lodash';
 import { getRelatedFiguresData, getRelatedResourcesData } from 'utils';
 
 const PaperCardStyled = styled.div`
@@ -135,21 +138,32 @@ class ComparisonCard extends Component {
                         )}
                         {this.props.comparison.description && (
                             <div>
-                                <small className="text-muted">{this.props.comparison.description}</small>
+                                <small className="text-muted">{truncate(this.props.comparison.description, { length: 150 })}</small>
                             </div>
                         )}
+                        <div className="d-block d-md-none mt-1">
+                            <RelativeBreadcrumbs researchField={this.props.comparison.researchField} />
+                        </div>
                     </Col>
 
-                    <Col md={3} style={{ marginTop: '8px' }}>
-                        {!this.state.openBox &&
-                            this.state.relatedFigures.slice(0, 2).map(url => (
-                                <ResourceItem key={url.figureId} style={{ float: 'right', padding: '3px' }}>
-                                    <div>
-                                        <Img src={url.src} alt={url.alt} />
-                                    </div>
-                                </ResourceItem>
-                            ))}
-                    </Col>
+                    <div className="col-md-3 text-right d-flex align-items-end" style={{ flexDirection: 'column' }}>
+                        <div style={{ flex: 1 }}>
+                            <div className="d-none d-md-block">
+                                <RelativeBreadcrumbs researchField={this.props.comparison.researchField} />
+                            </div>
+                            <div style={{ marginTop: '8px' }}>
+                                {!this.state.openBox &&
+                                    this.state.relatedFigures.slice(0, 2).map(url => (
+                                        <ResourceItem key={url.figureId} style={{ float: 'right', padding: '3px' }}>
+                                            <div>
+                                                <Img src={url.src} alt={url.alt} />
+                                            </div>
+                                        </ResourceItem>
+                                    ))}
+                            </div>
+                        </div>
+                        <UserAvatar userId={this.props.comparison.created_by} />
+                    </div>
                 </Row>
                 <Collapse isOpen={this.state.openBox}>
                     <hr />
@@ -203,7 +217,12 @@ ComparisonCard.propTypes = {
         reference: PropTypes.string,
         created_at: PropTypes.string,
         resources: PropTypes.array,
-        figures: PropTypes.array
+        figures: PropTypes.array,
+        researchField: PropTypes.shape({
+            id: PropTypes.string.isRequired,
+            label: PropTypes.string
+        }),
+        created_by: PropTypes.string
     }).isRequired,
     rounded: PropTypes.string,
     loadResources: PropTypes.bool
