@@ -85,7 +85,7 @@ const TableCellValueCreate = ({ isVisible, contributionId, propertyId, isEmptyCe
     return (
         <>
             {!isCreating && isVisible && (
-                <div className={isEmptyCell ? 'h-100' : ''} onDoubleClick={() => setIsCreating(true)}>
+                <div className={isEmptyCell ? 'h-100' : ''} role="button" tabIndex="0" onDoubleClick={() => setIsCreating(true)}>
                     <CreateButtonContainer className="create-button">
                         <StatementOptionButton title="Add value" icon={faPlus} action={() => setIsCreating(true)} />
                     </CreateButtonContainer>
