@@ -1,9 +1,9 @@
-import { Component } from 'react';
-import { Button, Modal, ModalHeader, ModalBody } from 'reactstrap';
-import styled from 'styled-components';
-import { FontAwesomeIcon as Icon } from '@fortawesome/react-fontawesome';
 import { faPlayCircle } from '@fortawesome/free-regular-svg-icons';
+import { FontAwesomeIcon as Icon } from '@fortawesome/react-fontawesome';
 import Tippy from '@tippyjs/react';
+import { useState } from 'react';
+import { Button, Modal, ModalBody, ModalHeader } from 'reactstrap';
+import styled from 'styled-components';
 
 const PlayButton = styled(Button)`
     font-size: 0.9rem !important;
@@ -23,58 +23,37 @@ const PlayButton = styled(Button)`
     }
 `;
 
-const VideoContainer = styled.div`
-    position: relative;
-    padding-top: 56.25%;
-`;
+const Video = () => {
+    const [isOpen, setIsOpen] = useState(false);
 
-const IframeFullWidth = styled.iframe`
-    position: absolute;
-    top: 0;
-    left: 0;
-    width: 100%;
-    height: 100%;
-    border: 0;
-`;
+    return (
+        <div>
+            <PlayButton color="link" className="p-0 m-0" onClick={() => setIsOpen(true)}>
+                <Tippy content="Play introduction video" offset={[0, 50]}>
+                    <span>
+                        <Icon icon={faPlayCircle} />
+                    </span>
+                </Tippy>
+            </PlayButton>
 
-export default class Video extends Component {
-    state = {
-        showDialog: false
-    };
+            <Modal isOpen={isOpen} toggle={() => setIsOpen(v => !v)} size="lg">
+                <ModalHeader toggle={() => setIsOpen(v => !v)}>ORKG introduction video</ModalHeader>
+                <ModalBody>
+                    <div class="embed-responsive embed-responsive-16by9">
+                        <iframe
+                            title="Explanation and introduction video of the ORKG"
+                            scrolling="no"
+                            frameBorder="0"
+                            src="//av.tib.eu/player/16120"
+                            allow="accelerometer; autoplay; encrypted-media; gyroscope; picture-in-picture"
+                            allowFullScreen={true}
+                            class="embed-responsive-item"
+                        />
+                    </div>
+                </ModalBody>
+            </Modal>
+        </div>
+    );
+};
 
-    toggleDialog = type => {
-        this.setState(prevState => ({
-            showDialog: !prevState.showDialog
-        }));
-    };
-
-    render() {
-        return (
-            <div>
-                <PlayButton color="link" className="p-0 m-0" onClick={this.toggleDialog}>
-                    <Tippy content="Play introduction video" offset={[0, 50]}>
-                        <span>
-                            <Icon icon={faPlayCircle} />
-                        </span>
-                    </Tippy>
-                </PlayButton>
-
-                <Modal isOpen={this.state.showDialog} toggle={this.toggleDialog} size="lg">
-                    <ModalHeader toggle={this.toggleDialog}>ORKG introduction video</ModalHeader>
-                    <ModalBody>
-                        <VideoContainer>
-                            <IframeFullWidth
-                                title="Explanation and introduction video of the ORKG"
-                                scrolling="no"
-                                frameBorder="0"
-                                src="//av.tib.eu/player/16120"
-                                allow="accelerometer; autoplay; encrypted-media; gyroscope; picture-in-picture"
-                                allowFullScreen={true}
-                            />
-                        </VideoContainer>
-                    </ModalBody>
-                </Modal>
-            </div>
-        );
-    }
-}
+export default Video;
