@@ -1,10 +1,8 @@
 import { useState, useEffect, useCallback } from 'react';
 import { getResearchFieldsByResearchProblemId } from 'services/backend/problems';
-import { useParams } from 'react-router-dom';
 
-function useResearchProblemResearchFields() {
+function useResearchProblemResearchFields({ researchProblemId }) {
     const [researchFields, setResearchFields] = useState([]);
-    const { researchProblemId } = useParams();
     const [isLoadingResearchFields, setIsLoadingResearchFields] = useState(true);
 
     const loadResearchProblemResearchFields = useCallback(rpId => {
@@ -25,4 +23,5 @@ function useResearchProblemResearchFields() {
     }, [researchProblemId, loadResearchProblemResearchFields]);
     return [researchFields, isLoadingResearchFields];
 }
+
 export default useResearchProblemResearchFields;
