@@ -5,7 +5,7 @@ import contributionEditorPreview from 'assets/img/AddComparison/contribution-edi
 import csvImportPreview from 'assets/img/AddComparison/csv-import-preview.png';
 import ROUTES from 'constants/routes';
 import { Link } from 'react-router-dom';
-import { Alert, Button, Col, Container, Row } from 'reactstrap';
+import { Button, Col, Container, Row } from 'reactstrap';
 
 const AddComparison = () => {
     document.title = 'Add comparison - Open Research Knowledge Graph';
@@ -15,17 +15,34 @@ const AddComparison = () => {
             <Container>
                 <h1 className="h4 mt-4 mb-4">Add comparison</h1>
             </Container>
-            <Container className="box rounded pt-4 pb-4 pl-5 pr-5">
-                <Alert color="info" fade={false}>
-                    Comparisons in ORKG provide an overview of state-of-the-art literature for a particular topic. Comparisons are dynamic and FAIR. A
-                    comparison is created from contributions,{' '}
-                    <a href="https://www.orkg.org/orkg/comparison/R44930" target="_blank" rel="noreferrer">
-                        view example of comparison <Icon icon={faExternalLinkAlt} />
-                    </a>
-                    . To create your own comparisons in ORKG, you can either import existing data (via CSV import) or start from scratch by adding
-                    your own contributions. This page guides you in creating new comparisons.
-                </Alert>
-                <Row className="mt-5">
+            <Container className="box rounded py-4 px-5">
+                <Row>
+                    <Col md="4">
+                        <div class="embed-responsive embed-responsive-16by9">
+                            <iframe
+                                title="How to make an ORKG comparison"
+                                scrolling="no"
+                                frameBorder="0"
+                                src="//av.tib.eu/player/51996"
+                                allow="accelerometer; autoplay; encrypted-media; gyroscope; picture-in-picture"
+                                allowFullScreen={true}
+                                class="embed-responsive-item"
+                            />
+                        </div>
+                    </Col>
+                    <Col md="8">
+                        Comparisons in ORKG provide an overview of state-of-the-art literature for a particular topic. Comparisons are dynamic and
+                        FAIR. A comparison is created from contributions,{' '}
+                        <a href="https://www.orkg.org/orkg/comparison/R44930" target="_blank" rel="noreferrer">
+                            view example of comparison <Icon icon={faExternalLinkAlt} />
+                        </a>
+                        . To create your own comparisons in ORKG, you can either import existing data (via CSV import) or start from scratch by adding
+                        your own contributions. This page guides you in creating new comparisons.
+                    </Col>
+                </Row>
+            </Container>
+            <Container className="box rounded py-4 px-5 mt-4">
+                <Row className="mt-2">
                     <Col className="border-right">
                         <h2 className="h4">1. Existing data</h2>
                         <div className="d-flex justify-content-center" style={{ minHeight: 160 }}>
@@ -35,7 +52,7 @@ const AddComparison = () => {
                             In case you have existing data, you can import this via the CSV import tool. This is especially helpful if you already
                             have a large file in which related work is compared.{' '}
                         </p>
-                        <Button color="primary" tag={Link} to={ROUTES.CSV_IMPORT} target="_blank">
+                        <Button color="primary" tag={Link} to={ROUTES.CSV_IMPORT}>
                             Go to CSV import tool
                         </Button>
                     </Col>
@@ -53,7 +70,7 @@ const AddComparison = () => {
                             If you don’t have existing data, go to the contribution editor to add contributions that will be used in the comparison.
                             After creating contributions, you can create a comparisons.
                         </p>
-                        <Button color="primary" tag={Link} to={ROUTES.CONTRIBUTION_EDITOR} target="_blank">
+                        <Button color="primary" tag={Link} to={ROUTES.CONTRIBUTION_EDITOR}>
                             Go to contribution editor
                         </Button>
                     </Col>
