@@ -128,7 +128,12 @@ const Papers = () => {
                     {paperResources.length > 0 &&
                         paperResources.map(paper => {
                             const paperCardData = statements.find(({ id }) => id === paper.id);
-                            return <PaperCardDynamic paper={{ title: paper.label, id: paper.id, paperData: paperCardData }} key={`pc${paper.id}`} />;
+                            return (
+                                <PaperCardDynamic
+                                    paper={{ title: paper.label, id: paper.id, paperData: paperCardData, created_by: paper.created_by }}
+                                    key={`pc${paper.id}`}
+                                />
+                            );
                         })}
                     {totalElements === 0 && !isNextPageLoading && (
                         <ListGroupItem tag="div" className="text-center p-4">
