@@ -9,12 +9,24 @@ export const getResearchFieldsByResearchProblemId = problemId => {
 };
 
 export const getContributorsByResearchProblemId = ({ id, page = 0, items = 9999 }) => {
-    const params = queryString.stringify({ page: page, size: items });
+    const params = queryString.stringify(
+        { page: page, size: items },
+        {
+            skipNull: true,
+            skipEmptyString: true
+        }
+    );
     return submitGetRequest(`${problemsUrl}${encodeURIComponent(id)}/users?${params}`);
 };
 
 export const getAuthorsByResearchProblemId = ({ id, page = 0, items = 9999 }) => {
-    const params = queryString.stringify({ page: page, size: items });
+    const params = queryString.stringify(
+        { page: page, size: items },
+        {
+            skipNull: true,
+            skipEmptyString: true
+        }
+    );
     return submitGetRequest(`${problemsUrl}${encodeURIComponent(id)}/authors?${params}`);
 };
 
