@@ -57,7 +57,13 @@ export const updateStatements = (statementIds, { subject_id = null, predicate_id
 
 export const getAllStatements = ({ page = 0, items: size = 9999, sortBy = 'created_at', desc = true }) => {
     const sort = `${sortBy},${desc ? 'desc' : 'asc'}`;
-    const params = queryString.stringify({ page, size, sort });
+    const params = queryString.stringify(
+        { page, size, sort },
+        {
+            skipNull: true,
+            skipEmptyString: true
+        }
+    );
 
     return submitGetRequest(`${statementsUrl}?${params}`).then(res => res.content);
 };
@@ -72,7 +78,13 @@ export const deleteStatementsByIds = ids => {
 
 export const getStatementsBySubject = ({ id, page = 0, items: size = 9999, sortBy = 'created_at', desc = true }) => {
     const sort = `${sortBy},${desc ? 'desc' : 'asc'}`;
-    const params = queryString.stringify({ page, size, sort });
+    const params = queryString.stringify(
+        { page, size, sort },
+        {
+            skipNull: true,
+            skipEmptyString: true
+        }
+    );
 
     return submitGetRequest(`${statementsUrl}subject/${encodeURIComponent(id)}/?${params}`).then(res => res.content);
 };
@@ -83,7 +95,13 @@ export const getStatementsBundleBySubject = ({ id }) => {
 
 export const getStatementsBySubjects = ({ ids, page = 0, items: size = 9999, sortBy = 'created_at', desc = true }) => {
     const sort = `${sortBy},${desc ? 'desc' : 'asc'}`;
-    const params = queryString.stringify({ ids: ids.join(), page, size, sort });
+    const params = queryString.stringify(
+        { ids: ids.join(), page, size, sort },
+        {
+            skipNull: true,
+            skipEmptyString: true
+        }
+    );
     return submitGetRequest(`${statementsUrl}subjects/?${params}`).then(res =>
         res.map(subjectStatements => ({
             ...subjectStatements,
@@ -94,7 +112,13 @@ export const getStatementsBySubjects = ({ ids, page = 0, items: size = 9999, sor
 
 export const getStatementsByObject = async ({ id, page = 0, items: size = 9999, sortBy = 'created_at', desc = true }) => {
     const sort = `${sortBy},${desc ? 'desc' : 'asc'}`;
-    const params = queryString.stringify({ page, size, sort });
+    const params = queryString.stringify(
+        { page, size, sort },
+        {
+            skipNull: true,
+            skipEmptyString: true
+        }
+    );
 
     const statements = await submitGetRequest(`${statementsUrl}object/${encodeURIComponent(id)}/?${params}`).then(res => res.content);
 
@@ -103,14 +127,26 @@ export const getStatementsByObject = async ({ id, page = 0, items: size = 9999, 
 
 export const getStatementsByPredicate = ({ id, page = 0, items: size = 9999, sortBy = 'created_at', desc = true, returnContent = true }) => {
     const sort = `${sortBy},${desc ? 'desc' : 'asc'}`;
-    const params = queryString.stringify({ page, size, sort });
+    const params = queryString.stringify(
+        { page, size, sort },
+        {
+            skipNull: true,
+            skipEmptyString: true
+        }
+    );
 
     return submitGetRequest(`${statementsUrl}predicate/${encodeURIComponent(id)}/?${params}`).then(res => (returnContent ? res.content : res));
 };
 
 export const getStatementsBySubjectAndPredicate = ({ subjectId, predicateId, page = 0, items: size = 9999, sortBy = 'created_at', desc = true }) => {
     const sort = `${sortBy},${desc ? 'desc' : 'asc'}`;
-    const params = queryString.stringify({ page, size, sort });
+    const params = queryString.stringify(
+        { page, size, sort },
+        {
+            skipNull: true,
+            skipEmptyString: true
+        }
+    );
 
     return submitGetRequest(`${statementsUrl}subject/${subjectId}/predicate/${predicateId}/?${params}`).then(res => res.content);
 };
@@ -125,7 +161,13 @@ export const getStatementsByObjectAndPredicate = ({
     returnContent = true
 }) => {
     const sort = `${sortBy},${desc ? 'desc' : 'asc'}`;
-    const params = queryString.stringify({ page, size, sort });
+    const params = queryString.stringify(
+        { page, size, sort },
+        {
+            skipNull: true,
+            skipEmptyString: true
+        }
+    );
 
     return submitGetRequest(`${statementsUrl}object/${objectId}/predicate/${predicateId}/?${params}`).then(res =>
         returnContent ? res.content : res
